Lunar Surface Density

Firstly, the ExosphereSimulations.jl package has to be loaded to be able to use its entire functionality:

include(joinpath(path, to, repository, "src", "ExosphereSimulations.jl"))
using .ExosphereSimulations

Now, both the base functions, as well as the sub-module LunarSurfaceDensity with all of its functionality are available. The usual workflow for setting up and running a simulation is based on first defining the problem using the LSDProblem constructor and then solving the problem using the solve function.

prob = LSDProblem(grid, physics, opt)
sol = solve(prob)

While prob can also be created without passing the solver options opt to the constructor, both the numerical grid grid and the physical setup physics have to be known. The implemented numerical grids are:

StructuredFullSGrid,

StructuredHalfSGrid,

FSpiralFullSGrid,

FSpiralHalfSGrid,

HEALPixFullSGrid.

The physical parameters use a specific constructor called LSDPhysics, which holds the necessary physical description of the problem based on input parameters. A more detailed list can be found on the input page, which lists input parameters for various different species alongside a description and references. The general construction of the physics object is:

physics = LSDPhysics(amu, source_global, source_sunside, photorate, adheat, recycling, vdist)

Velocity Distribution

In the current version, the variable vdist has to be of type Symbol referring to one of the implemented velocity distributions. The allowed options are:

:Maxwell

:MaxwellBoltzmannFlux

Finally, the solver options can be set manually, if the user should want to make changes to the default values. The constructor in question is LSDOptions and can be created empty

opt = LSDOptions()

to access all the default parameters, or using a keyword-value pair like

opt = LSDOptions(; N = 2_500)

which is adjusting the number of Monte-Carlo steps N. The available keywords can be found in the LSDOptions docstring.

Finally, now that the problem object prob has been created, we can pass it to the solver by

sol = solve(prob)

which creates a solution object of type LSDSolution, containing both the original problem prob, as well as the results of the simulation. Most importantly, it contains the predictions for the surface number densities $n_0$ in cm$^{-3}$ as sol.n0.

API

ExosphereSimulations.LunarSurfaceDensity — Module

Julia Package for simulating the surface number densities of the lunar surface-bounded exosphere.

ExosphereSimulations.LunarSurfaceDensity.LSDOptions — Type

LSDOptions(;
    N           ::  Int64     = 1000,
    vdist_N     ::  Int64     = 100_000,
    vdist_atol  ::  Float64   = 1e-1,
    photo_thr   ::  Float64   = 1e-3,
    dt          ::  Float64   = MOON_DAY / 360,
    ctrap_thr   ::  Float64   = 10.0,
    h0          ::  Float64   = 1.0,
    h0_N        ::  Int64     = 10,
)

Custom container for all required solver options for the lunar surface density study. The description of the solver options is shown below:

Option	Description	Effect

N	Number of Monte-Carlo steps	Simulation time, result accuracy

vdist_N	Number of sample points for MBF velocity distribution	MBF velocity, MBF calc. time

vdist_atol	Absoltue tolerance of the MBF velocity calculation	MBF calc. accuracy & time

photo_thr	Threshold of photo-processes	Simulation time, result accuracy

dt	Time step for surface interaction calculations	Simulation time, ad-/desorption accuracy

ctrap_thr	Cold-trap treshold

h0	Surface height

h0_N	Number of surface height intervalls	Simulation time, result accuracy

Notes

vdist_N & vdist_atol are only relevant if the used velocity distribution is a Maxwell-Boltzmann Flux distribution

vdist_N increases the maximum possible velocity

h0_N gets more relevant for heavier particles and bigger h0 values

ExosphereSimulations.LunarSurfaceDensity.LSDPhysics — Type

LSDPhysics(
    amu             ::  Float64, 
    source_global   ::  Float64, 
    source_sunside  ::  Float64, 
    photorate       ::  Float64, 
    adheat          ::  Float64
    recycling       ::  Float64, 
    vdist           ::  Symbol)

Custom container for all required physical input parameters for the lunar surface density study.

ExosphereSimulations.LunarSurfaceDensity.LSDProblem — Type

LSDProblem{T<:SGrid2D}(
    grid    ::  T,
    physics ::  LSDPhysics,
    opt     ::  LSDOptions,
    phi     ::  Vector{Float64},
    T       ::  Vector{Float64},
    tres    ::  Vector{Float64},
    vdist   ::  VelocityDistribution,
)
LSDProblem(grid::SGrid2D, physics::LSDPhysics, opt::LSDOptions)
LSDProblem(grid::SGrid2D, physics::LSDPhysics)

Custom container of all necessary information to describe the lunar surface density simulation problem. It is not recommended to directly construct a LSDProblem, but to use one of the additionally defined outer constructors.

LSDProblem can be constructed without passing a opt to the constructor, at which the constructor will use the default LSDOptions values.

Field	Description

grid	Contains geometry and coordinates of the numerical grid

physics	Contains physical input parameters for the simulation

opt	Contains solver options

phi	Solar incidence angles on the computational grid

T	Surface temperatures on the computational grid

tres	Residence time of the particle inside of the lunar regolith

vdist	ExosphereSimulations velocity distribution struct

ExosphereSimulations.LunarSurfaceDensity.LSDSolution — Type

LSDSolution(prob::LSDProblem, n0::Vector{FLoat64}, los::Vector{Float64})

Custom solution struct for the solution of the lunar surface density simulation. It holds the initial problem definition prob, as well as the resulting surface density n0 over the defined numerical grid in [cm-3] and the fractions of lost particles lost in [s-1].

Description of lost vector

Description

lost[1]	Fraction of particles lost through thermal escape

lost[2]	Fraction of particles lost through ionization processes

lost[3]	Fraction of particles lost in cold-traps

The surface density can be directly accessed by indexing the solution object with normal brackets:

julia> sol = solve(prob);

julia> sol(12)
40231.123415

ExosphereSimulations.LunarSurfaceDensity.savesol — Function

savesol(name::String, sol::LSDSolution)

Saves the solutions sol longitudes, latitudes, and surface number densities in a .csv file specified in name. The coordinates are taken directly from the underlying numerical grid.

ExosphereSimulations.LunarSurfaceDensity.gridsavesol — Function

gridsavesol(name::String, sol::LSDSolution, glngs::Vector{T}, glats::Vector{S}) where {T <: Real, S <: Real})

Saves the solutions sol longitudes, latitudes, and surface number densities in a .csv file specified in name. The coordinates are a grid created using the inputs glngs and glats, which are the longitudinal and latitudinal sample points.

ExosphereSimulations.LunarSurfaceDensity.solve — Function

solve(prob::LSDProblem; logstep=10)

Main simulation function of the lunar surface density simulation module. Requires prior definition of the module specific problem struct prob of type LSDProblem. The input logstep sets the intervall for logged information to be printed in the terminal during the simulation.

The function features a parallelized simulation call that splits the Monte-Carlo steps onto the available workers.

ExosphereSimulations.LunarSurfaceDensity.path — Function

path(prob::LSDProblem, idx::Int64, w::Real)

Calculate the path of a particle, initially launched from the grid element with index idx with a weight of w. Loops until the particle is lost. Returns particle density along the path and the lost fraction vector.

ExosphereSimulations.LunarSurfaceDensity._height_time — Function

_height_time(v::Real, h=1e0; N=1)

Internal function to compute the time a particle with vertical velocity v in meters per second requires to reach h meter above ground. The additional variable N gives the number of parts the height is split into while calculating the total travel time.

ExosphereSimulations.LunarSurfaceDensity._particle_flight! — Function

_particle_flight!(prob::LSDProblem, N0::AbstractVector, idx::Int64, coords::Tuple, w::Real, w0::Real)

Internal function to handle the ballistic flight of the particle launched at coordinates given in coords, at the grid element of index idx, with a current weight of w and a initial weight of w0. Fills the entries of the resulting particle density vector N0 depending on the launch and landing position and velocity. Returns the landing grid element index, the exact landing coordinates, and the new, updated weight.

ExosphereSimulations.LunarSurfaceDensity._surface_interaction — Function

_surface_interaction(prob::LSDProblem, idx::Int64, coords::Tuple)

Internal function to handle the surface interaction. Dynamically adjusts the longitude with respect to the time the particle has spend inside of the surface. Returns the index of the element from where the particle is released, along side its exact coordinates.