Solar Angle

ExosphereSimulations.solar_angle — Function

solar_angle(lng::Real, lat::Real) -> Float64
solar_angle(grid::SGrid2D) -> Vector{Float64}

Calculate the solar incidence angle based in selenocentric coordinates given as lng and lat in radians.

ExosphereSimulations.solar_angle_deg — Function

solar_angle_deg(lng::Real, lat::Real) -> Float64
solar_angle_deg(grid::SGrid2D) -> Vector{Float64}

Specific version of solar_angle(lng, lat) for both the input and output angles given and returned in degree.