Velocity Distribution

Maxwell Velocity Distribution

ExosphereSimulations.MaxwellVelocityDistribution — Type

MaxwellVelocityDistribution{T<:Real}(amu::T)

Struct for the Maxwell velocity distribution holding the particle atomic mass unit amu.

The object itself can be used to draw a random sample vector from this velocity distribution. As input it requires only the temperature of the thermal equilibrium T:

julia> max = MaxwellVelocityDistribution(4.0026);

julia> max(250)
3-element Vector{Float64}:
 -340.8406689399358
 -563.3849738101032
  105.53990594265329

Maxwell-Boltzmann Flux Velocity Distribution

ExosphereSimulations.MaxwellBoltzmannFluxVelocityDistribution — Type

MaxwellBoltzmannFluxVelocityDistribution{T<:Real}(itp::Interpolations.Extrapolation, amu::T)
MaxwellBoltzmannFluxVelocityDistribution(amu::Real; N=10_000, atol=1e-1)

Custom struct used to draw three-dimensional velocity vectors from a Maxwell-Boltzmann flux velocity distribution. Holds an interpolation object itp with precomputed values of the distribution, and the atomic weight amu of the respective particle for which the velocity is drawn.

The recommended use is to create the struct using the outer constructor, which creates the interpolation object. Afterwards, the struct can be used directly to sample the distribution requiring the temperature as its only input.

julia> mbf = MaxwellBoltzmannFluxVelocityDistribution(4.0026);

julia> mbf(250)
3-element Vector{Float64}:
 -102.65571182106781
  892.2770113419351
  683.0078350151742